backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Cache
redis==5.0.1
aioredis==2.0.1

# TTS & Audio
azure-cognitiveservices-speech==1.34.0
pydub==0.25.1
ffmpeg-python==0.2.0

# AI & NLP
openai==1.10.0
langchain==0.1.4
transformers==4.37.2

# WebSocket
websockets==12.0

# Task Queue
celery==5.3.6
flower==2.0.1

# Monitoring
prometheus-client==0.19.0

# Logging
python-json-logger==2.0.7

# Utils
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.1

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
